refactor DB into inner struct, use u64 instead of usize

+++ src/routes.rs
@@ -84,7 +84,7 @@ pub async fn upload_file(
         // push to entries
         entries.push(DBEntry::new(
             filename,
-            file_size,
+            file_size as u64,
             random(5).collect::<String>(),
         ));
     }
